minor grammar/punctuation; adding a link to RaspiBolt
This PR is mostly for improving some minor grammar and punctuation errors.

I also added a link to the RaspiBolt guide, on which I have second thoughts now, as perhaps the idea of the link collection was more to present one link for each type of a node (DIY, DIY with premium options, "out of the box" on a Raspberry, and a "out of the box" on a more powerful device). So maybe then an additional Raspberry DIY guide is redundant here.